adding questions qcov16 and qcov16b to common pool
