+++ src/editor-example-generators/italy/inf-weekly.ts
@@ -1,10 +1,10 @@
-import { Survey, SurveyItem, SurveyGroupItem } from "survey-engine/lib/data_types";
+import { Survey, SurveyGroupItem, SurveyItem } from "survey-engine/lib/data_types";
 import { ItemEditor } from "../../editor-engine/survey-editor/item-editor";
 import { SurveyEditor } from "../../editor-engine/survey-editor/survey-editor";
-import { WeeklyQuestions as CommonPoolWeekly } from "../common_question_pool/influenzanet-weekly";
-import { initLikertScaleItem, initMatrixQuestion, initDropdownGroup, initMultipleChoiceGroup, initSingleChoiceGroup, ResponseRowCell } from "../../editor-engine/utils/question-type-generator";
+import { initDropdownGroup, initLikertScaleItem, initMultipleChoiceGroup, initSingleChoiceGroup } from "../../editor-engine/utils/question-type-generator";
 import { expWithArgs, generateHelpGroupComponent, generateLocStrings, generateTitleComponent } from "../../editor-engine/utils/simple-generators";
-import { likertScaleKey, matrixKey, multipleChoiceKey, responseGroupKey, singleChoiceKey } from "../common_question_pool/key-definitions";
+import { WeeklyQuestions as CommonPoolWeekly } from "../common_question_pool/influenzanet-weekly";
+import { likertScaleKey, multipleChoiceKey, responseGroupKey, singleChoiceKey } from "../common_question_pool/key-definitions";
 
 const weekly = (): Survey | undefined => {
     const surveyKey = 'weekly';
@@ -136,6 +136,14 @@ const weekly = (): Survey | undefined => {
     const Q_consFear = CommonPoolWeekly.consFear(hasSymptomGroupKey, Q_visitedMedicalService.key, Q_contactedMedicalService.key, Q_visitedNoMedicalService.key, true);
     survey.addExistingSurveyItem(Q_consFear, hasSymptomGroupKey);
 
+    // // Qcov16 test -----------------------------------------------------
+    const Q_SymptomImpliedCovidTest = CommonPoolWeekly.symptomImpliedCovidTest(hasSymptomGroupKey, true);
+    survey.addExistingSurveyItem(Q_SymptomImpliedCovidTest, hasSymptomGroupKey);
+
+    //Qcov16b test result
+    const Q_resultTest = CommonPoolWeekly.resultPCRTest(hasSymptomGroupKey, Q_SymptomImpliedCovidTest.key, true)
+    survey.addExistingSurveyItem(Q_resultTest, hasSymptomGroupKey);
+
     // // Q9 took medication --------------------------------------
     const Q_tookMedication = CommonPoolWeekly.tookMedication(hasSymptomGroupKey, true);
     survey.addExistingSurveyItem(Q_tookMedication, hasSymptomGroupKey);
@@ -173,10 +181,6 @@ const weekly = (): Survey | undefined => {
     const Q_durationTest = durationTest(rootKey, Q_covidTest.key, Q_reasonTest.key, true, "Q_BE_cov16e")
     survey.addExistingSurveyItem(Q_durationTest, rootKey);
 
-    //Qcov_BE_16b test result
-    const Q_resultTest = resultTest(rootKey, Q_covidTest.key, true, "Qcov_BE_16b")
-    survey.addExistingSurveyItem(Q_resultTest, rootKey);
-
     //Q_BE_cov16z duration untill test result
     const Q_durationTestResult = durationTestResult(rootKey, Q_covidTest.key, Q_resultTest.key, true, "Qcov_BE_16z")
     survey.addExistingSurveyItem(Q_durationTestResult, rootKey);
@@ -195,10 +199,6 @@ const weekly = (): Survey | undefined => {
     // // Q11 think cause of symptoms --------------------------------------
     const Q_causeOfSymptoms = CommonPoolWeekly.causeOfSymptoms(hasSymptomGroupKey, true);
     survey.addExistingSurveyItem(Q_causeOfSymptoms, hasSymptomGroupKey);
-
-    // // Qcov_BE_16 test -----------------------------------------------------
-    const Q_SymptomImpliedCovidTest = SymptomImpliedCovidTest(hasSymptomGroupKey, true, "Qcov_BE_16");
-    survey.addExistingSurveyItem(Q_SymptomImpliedCovidTest, hasSymptomGroupKey);
 
     // // Qcov_BE_7 Covid 19 habits change question ------------------------------------------------------
     const Q_covidHabits = covidHabitsChange(hasSymptomGroupKey, false, "Qcov_BE_7");
@@ -720,123 +720,6 @@ const durationTest = (parentKey: string, keycovidTest?: string, keyreasonTest?: 
 }
 
 /**
- * RESULT COVID-19 TEST: result COVID-19 test
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const resultTest = (parentKey: string, keycovidTest?: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Qcov_BE_16b'
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Heeft u de COVID-19 testresultaten reeds ontvangen?"],
-            ["fr-be", "Avez-vous déjà reçu les résultats du test de dépistage du coronavirus ?"],
-            ["de-be", "Haben Sie die COVID-19-Testergebnisse bereits erhalten?"],
-            ["en", "Have you received the results of your coronavirus test?"],
-        ]))
-    );
-
-    // CONDITION
-    if (keycovidTest) {
-        editor.setCondition(
-            expWithArgs('responseHasKeysAny', keycovidTest, [responseGroupKey, singleChoiceKey].join('.'), '0')
-        );
-    }
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Waarom vragen we dit?"],
-                    ["fr-be", "Pourquoi posons-nous cette question ?"],
-                    ["de-be", "Warum fragen wir das?"],
-                    ["en", "Why are we asking this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "We willen weten hoe COVID-19 zich verspreidt in de bevolking."],
-                    ["fr-be", "Nous voulons savoir comment le coronavirus se propage au sein de la population."],
-                    ["de-be", "Wir möchten wissen, wie COVID-19 sich in der Bevölkerung ausbreitet."],
-                    ["en", "We want to understand how the coronavirus is spreading within the population."],
-                ]),
-                //style: [{ key: 'variant', value: 'p' }],
-            },
-        ])
-    );
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    const rg_inner = initSingleChoiceGroup(singleChoiceKey, [
-        {
-            key: '1', role: 'option',
-            content: new Map([
-                ["nl-be", "Ja, positief voor COVID-19"],
-                ["fr-be", "Oui, positif au coronavirus"],
-                ["de-be", "Ja, positiv auf COVID-19"],
-                ["en", "Yes, the test is positive for coronavirus"],
-            ])
-        },
-        {
-            key: '2', role: 'option',
-            content: new Map([
-                ["nl-be", "Ja, negatief voor COVID-19"],
-                ["fr-be", "Oui, négatif au coronavirus"],
-                ["de-be", "Ja, negativ auf COVID-19"],
-                ["en", "Yes, the test is negative for coronavirus"],
-            ])
-        },
-        {
-            key: '3', role: 'option',
-            content: new Map([
-                ["nl-be", "Ja, niet-interpreteerbaar resultaat"],
-                ["fr-be", "Oui, résultat non interprétable"],
-                ["de-be", "Ja, nicht interpretierbares Ergebnis"],
-                ["en", "Yes, the results are inconclusive"],
-            ])
-        },
-        {
-            key: '4', role: 'option',
-            content: new Map([
-                ["nl-be", "Nee, ik heb nog geen testresultaat"],
-                ["fr-be", "Non, je n'ai pas encore reçu le résultat du test"],
-                ["de-be", "Nein, ich habe noch kein Testergebnis"],
-                ["en", "No, I have not yet received the test results"],
-            ])
-        },
-        {
-            key: '5', role: 'option',
-            content: new Map([
-                ["nl-be", "Dat wil ik niet aangeven"],
-                ["fr-be", "Je préfère ne pas le spécifier"],
-                ["de-be", "Das möchte ich nicht angeben"],
-                ["en", "I prefer not to specify"],
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
-
-/**
  * DURATION COVID-19 TEST RESULT: duration untill COVID-19 test result
  * TO DO: fix dropdown menu
  *
@@ -1010,854 +893,6 @@ const durationTestResult = (parentKey: string, keycovidTest?: string, keyresultT
 
     const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
     editor.addExistingResponseComponent(ddOptions, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
-
-/**
- * VISITED MEDICAL SERVICE
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const visitedMedicalService = (parentKey: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Q_BE_7';
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Heeft u medische hulp gezocht vanwege uw symptomen/klachten? En zo ja, waar?"],
-            ["fr-be", "Avez-vous consulté un médecin en raison de vos symptômes/plaintes ? Et si oui, où ?"],
-            ["de-be", "Haben Sie wegen Ihrer Symptome/Beschwerden ärztliche Hilfe aufgesucht? Und wenn ja, wo?"],
-            ["en", "Did you consult medical assistance because of your symptoms / complaints? If yes, where?"],
-        ]))
-    );
-
-    // CONDITION
-    // None
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Waarom vragen we dit?"],
-                    ["fr-be", "Pourquoi posons-nous cette question ?"],
-                    ["de-be", "Warum fragen wir das?"],
-                    ["en", "Why are we asking this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Om uit te zoeken welk percentage van mensen met bepaalde klachten medische hulp zoeken."],
-                    ["fr-be", "Pour connaître le pourcentage de personnes présentant certains symptômes qui consultent un médecin."],
-                    ["de-be", "Um auszusuchen, welcher Anteil von Menschen mit bestimmten Beschwerden ärztliche Hilfe aufsuchen."],
-                    ["en", "To determine the percentage of people with certain symptoms who consult a doctor."],
-                ]),
-                style: [{ key: 'variant', value: 'p' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Hoe moet ik deze vraag beantwoorden?"],
-                    ["fr-be", "Comment dois-je répondre à cette question ?"],
-                    ["de-be", "Wie soll ich diese Frage beantworten?"],
-                    ["en", "How should I answer this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Selecteer alle relevante vormen van medische hulp die u heeft bezocht. Wanneer u nog niet bent geweest maar wel een afspraak heeft gemaakt, selecteer dan de laaste optie."],
-                    ["fr-be", "Sélectionnez toutes les formes d'aide médicale pertinentes que vous avez consultées. Si vous n'avez pas encore été chez un médecin, mais que vous avez fixé un rendez-vous, sélectionnez la dernière option."],
-                    ["de-be", "Wählen Sie alle relevanten Formen medizinischer Hilfe aus, die Sie besucht haben. Wenn Sie noch nicht dort gewesen sind, aber einen Termin vereinbart haben, dann wählen Sie die letzte Alternative."],
-                    ["en", "Select all of the types of relevant medical assistance you consulted. If you have not yet been to a doctor, but you have made an appointment, select the last option."],
-                ]),
-            },
-        ])
-    );
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-2' }],
-        content: generateLocStrings(
-            new Map([
-                ['nl-be', "Meerdere antwoorden mogelijk"],
-                ["fr-be", "Plusieurs réponses sont possibles"],
-                ["de-be", "Mehrere Antworten möglich"],
-                ["en", "Multiple answers possible"],
-            ])),
-    }, rg?.key);
-    const rg_inner = initMultipleChoiceGroup(multipleChoiceKey, [
-        {
-            key: '0', role: 'option',
-            disabled: expWithArgs('responseHasOnlyKeysOtherThan', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '0'),
-            content: new Map([
-                ["nl-be", "Nee, ik heb geen medische hulp gezocht"],
-                ["fr-be", "Non, je n'ai pas consulté un médecin"],
-                ["de-be", "Nein, ich habe keine ärztliche Hilfe aufgesucht"],
-                ["en", "No, I did not consult a doctor"],
-            ])
-        },
-        {
-            key: '1', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '0', '5'),
-            content: new Map([
-                ["nl-be", "Ja, fysiek bij de huisarts of huisarts assistent"],
-                ["fr-be", "Oui, physiquement au cabinet du médecin ou de l'assistant"],
-                ["de-be", "Ja, physisch beim Hausarzt oder beim Assistenten des Hausarztes"],
-                ["en", "Yes, in person at the doctor or assistant's office "],
-            ])
-        },
-        {
-            key: '31', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '0', '5'),
-            content: new Map([
-                ["nl-be", "Ja, fysiek bij de spoedafdeling van het ziekenhuis"],
-                ["fr-be", "Oui, physiquement au service des urgences de l'hôpital"],
-                ["de-be", "Ja, physisch in der Notaufnahme des Krankenhauses"],
-                ["en", "Yes, in person at the hospital emergency department"],
-            ])
-        },
-        {
-            key: '32', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '0', '5'),
-            content: new Map([
-                ["nl-be", "Ja, fysiek bij de huisartsenwachtpost"],
-                ["fr-be", "Oui, physiquement au service de garde de médecine générale"],
-                ["de-be", "Ja, physisch beim Wachtposten des Hausarztes"],
-                ["en", "Yes, in person at the on-call general medical service "],
-            ])
-        },
-        {
-            key: '6', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '0', '5'),
-            content: new Map([
-                ["nl-be", "Ja, maar enkel telefonisch of via e-mail contact met huisdokter, receptioniste, verpleger/verpleegster"],
-                ["fr-be", "Oui, mais seulement par le biais d'un contact téléphonique ou d'un courrier électronique avec le médecin, la réceptionniste, l'infirmier/l'infirmière"],
-                ["de-be", "Ja, aber nur telefonisch oder über einem E-Mail-Kontakt mit Hausarzt, Empfangsdame, Krankenpfleger/Krankenschwester"],
-                ["en", "Yes, but only through telephone or email contact with the doctor, receptionist, nurse "],
-            ])
-        },
-        {
-            key: '2', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '0', '5'),
-            content: new Map([
-                ["nl-be", "Ja, ik ben opgenomen in het ziekenhuis"],
-                ["fr-be", "Oui, j'ai été admis(e) à l'hôpital"],
-                ["de-be", "Ja, ich wurde in das Krankenhaus aufgenommen"],
-                ["en", "Yes, I was admitted into the hospital"],
-            ])
-        },
-        {
-            key: '4', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '0', '5'),
-            content: new Map([
-                ["nl-be", "Ja, ik heb andere medische hulp gezocht"],
-                ["fr-be", "Oui, j'ai fait appel à une autre aide médicale"],
-                ["de-be", "Ja, ich habe eine andere ärztliche Hilfe aufgesucht"],
-                ["en", "Yes, I consulted another healthcare assistance"],
-            ])
-        },
-        {
-            key: '5', role: 'option',
-            disabled: expWithArgs('responseHasOnlyKeysOtherThan', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '5'),
-            content: new Map([
-                ["nl-be", "Nog niet, maar ik heb een afspraak gemaakt"],
-                ["fr-be", "Pas encore, mais j'ai pris un rendez-vous"],
-                ["de-be", "Noch nicht, aber ich habe eine Verabredung gemacht"],
-                ["en", "Not yet, but I have taken an appointment"],
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
-
-/**
- * WHEN VISITED MEDICAL SERVICE
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param keyVisitedMedicalServ: reference to quesiton if visited any medical service
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const visitedMedicalServiceWhen = (parentKey: string, keyVisitedMedicalServ: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Q_BE_7b';
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Hoe snel na de start van uw symptomen/klachten heeft u voor de EERSTE keer medische hulp gezocht?"],
-            ["fr-be", "Combien de temps après l'apparition de vos symptômes/plaintes avez-vous consulté un médecin pour la PREMIÈRE fois ?"],
-            ["de-be", "Wie schnell nach dem Beginn Ihrer Symptome/Beschwerden haben Sie zum ERSTEN Mal ärztliche Hilfe aufgesucht?"],
-            ["en", "How soon after your symptoms appeared did you first visit this medical service?"],
-        ]))
-    );
-
-    // CONDITION
-    editor.setCondition(
-        expWithArgs('responseHasOnlyKeysOtherThan', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '0', '5')
-    );
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Waarom vragen we dit?"],
-                    ["fr-be", "Pourquoi posons-nous cette question ?"],
-                    ["de-be", "Warum fragen wir das?"],
-                    ["en", "Why are we asking this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Om uit te zoeken hoe snel mensen met klachten worden gezien door een medische hulpverlener/specialist."],
-                    ["fr-be", "Pour savoir à quelle vitesse les personnes présentant des symptômes sont vues par un médecin/spécialiste/professionnel de la santé."],
-                    ["de-be", "Um festzustellen, wie schnell Menschen mit Beschwerden von einem Arzthelfer/Facharzt gesehen werden."],
-                    ["en", "To know how quickly people with symptoms are seen by a doctor/specialist/health professional."],
-                ]),
-                style: [{ key: 'variant', value: 'p' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Hoe moet ik deze vraag beantwoorden?"],
-                    ["fr-be", "Comment dois-je répondre à cette question ?"],
-                    ["de-be", "Wie soll ich diese Frage beantworten?"],
-                    ["en", "How should I answer this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Geef alleen het aantal dagen van het begin van de klachten tot uw EERSTE bezoek aan de desbetreffende medische hulpverlener/specialist."],
-                    ["fr-be", "Indiquez simplement le nombre de jours écoulés entre le début des symptômes et votre PREMIÈRE visite auprès du médecin/spécialiste/professionnel de la santé en question."],
-                    ["de-be", "Bitte geben Sie nur die Anzahl der Tage vom Beginn der Beschwerden bis zu Ihrem ERSTEN Besuch bei dem betreffenden Arzthelfer/Spezialisten an."],
-                    ["en", "Simply indicate how many days passed between the onset of symptoms and your FIRST visit to the doctor/specialist/health professional."],
-                ]),
-            },
-        ])
-    );
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    editor.addExistingResponseComponent({
-        role: 'text',
-        content: generateLocStrings(
-            new Map([
-                ['nl-be', 'Selecteer het juiste aantal dagen'],
-                ["fr-be", ""],
-                ["de-be", ""],
-                ["en", ""],
-            ])),
-    }, rg?.key);
-    const ddOptions: ResponseRowCell = {
-        key: 'col1', role: 'dropDownGroup', items: [
-            {
-                key: '0', role: 'option', content: new Map([
-                    ["nl-be", "Op dezelfde dag als de eerste klachten"],
-                    ["fr-be", "Le même jour que l’apparition des premiers symptômes"],
-                    ["de-be", "An demselben Tag wie die ersten Beschwerden"],
-                    ["en", "Same day"],
-                ]),
-            },
-            {
-                key: '1', role: 'option', content: new Map([
-                    ["nl-be", "1 dag"],
-                    ["fr-be", "1 jour"],
-                    ["de-be", "1 Tag"],
-                    ["en", "1 day"],
-                ]),
-            },
-            {
-                key: '2', role: 'option', content: new Map([
-                    ["nl-be", "2 dagen"],
-                    ["fr-be", " 2 jours"],
-                    ["de-be", "2 Tage"],
-                    ["en", "2 days"],
-
-                ]),
-            },
-            {
-                key: '3', role: 'option', content: new Map([
-                    ["nl-be", "3 dagen"],
-                    ["fr-be", "3 jours"],
-                    ["de-be", "3 Tage"],
-                    ["en", "3 days"],
-                ]),
-            },
-            {
-                key: '4', role: 'option', content: new Map([
-                    ["nl-be", "4 dagen"],
-                    ["fr-be", "4 jours"],
-                    ["de-be", "4 Tage"],
-                    ["en", "4 days"],
-                ]),
-            },
-            {
-                key: '5', role: 'option', content: new Map([
-                    ["nl-be", "5 dagen"],
-                    ["fr-be", "5 jours"],
-                    ["de-be", "5 Tage"],
-                    ["en", "5 days"],
-                ]),
-            },
-            {
-                key: '6', role: 'option', content: new Map([
-                    ["nl-be", "6 dagen"],
-                    ["fr-be", "6 jours"],
-                    ["de-be", "6 Tage"],
-                    ["en", "6 days"],
-                ]),
-            },
-            {
-                key: '7', role: 'option', content: new Map([
-                    ["nl-be", "7 dagen"],
-                    ["fr-be", "7 jours"],
-                    ["de-be", "7 Tage"],
-                    ["en", "7 days"],
-                ]),
-            },
-            {
-                key: '8', role: 'option', content: new Map([
-                    ["nl-be", "8 dagen"],
-                    ["fr-be", "8 jours"],
-                    ["de-be", "8 Tage"],
-                    ["en", "8 days"],
-                ]),
-            },
-            {
-                key: '9', role: 'option', content: new Map([
-                    ["nl-be", "9 dagen"],
-                    ["fr-be", "9 jours"],
-                    ["de-be", "9 Tage"],
-                    ["en", "9 days"],
-                ]),
-            },
-            {
-                key: '10', role: 'option', content: new Map([
-                    ["nl-be", "10 dagen"],
-                    ["fr-be", "10 jours"],
-                    ["de-be", "10 Tage"],
-                    ["en", "10 days"],
-                ]),
-            },
-            {
-                key: '11', role: 'option', content: new Map([
-                    ["nl-be", "11 dagen"],
-                    ["fr-be", "11 jours"],
-                    ["de-be", "11 Tage"],
-                    ["en", "11 days"],
-                ]),
-            },
-            {
-                key: '12', role: 'option', content: new Map([
-                    ["nl-be", "12 dagen"],
-                    ["fr-be", "12 jours"],
-                    ["de-be", "12 Tage"],
-                    ["en", "12 days"],
-                ]),
-            },
-            {
-                key: '13', role: 'option', content: new Map([
-                    ["nl-be", "13 dagen"],
-                    ["fr-be", "13 jours"],
-                    ["de-be", "13 Tage"],
-                    ["en", "13 days"],
-                ]),
-            },
-            {
-                key: '14', role: 'option', content: new Map([
-                    ["nl-be", "14 dagen"],
-                    ["fr-be", "14 jours"],
-                    ["de-be", "14 Tage"],
-                    ["en", "14 days"],
-                ]),
-            },
-            {
-                key: '15', role: 'option', content: new Map([
-                    ["nl-be", "meer dan 14 dagen"],
-                    ["fr-be", "+ de 14 jours"],
-                    ["de-be", "14+ Tage"],
-                    ["en", "14+ days"],
-                ]),
-            },
-            {
-                key: '16', role: 'option', content: new Map([
-                    ["nl-be", "Ik weet het niet (meer)"],
-                    ["fr-be", "Je ne sais pas (plus)"],
-                    ["de-be", "Ich weiß es nicht (mehr)"],
-                    ["en", "I don’t know/can’t remember"],
-                ]),
-            },
-        ]
-    };
-
-    const rg_inner = initMatrixQuestion(matrixKey, [
-        {
-            key: 'header', role: 'headerRow', cells: [
-                {
-                    key: 'col0', role: 'text', content: new Map([
-                        ["nl-be", "Medische hulp"],
-                        ["fr-be", "Aide médicale"],
-                        ["de-be", "ärztliche Hilfe"],
-                        ["en", "Healthcare assistance"],
-                    ]),
-                },
-                {
-                    key: 'col1', role: 'text'
-                },
-            ]
-        },
-        {
-            key: 'r1', role: 'responseRow', cells: [
-                {
-                    key: 'col0', role: 'label', content: new Map([
-                        ["nl-be", "Fysiek bij de huisarts of huisarts assistent"],
-                        ["fr-be", "Physiquement au cabinet du médecin ou de l'assistant"],
-                        ["de-be", "Physisch beim Hausarzt oder beim Assistenten des Hausarztes"],
-                        ["en", "In person at the doctor or assistant's office "],
-                    ]),
-                },
-                { ...ddOptions }
-            ],
-            displayCondition: expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '1')
-        },
-        {
-            key: 'r31', role: 'responseRow', cells: [
-                {
-                    key: 'col0', role: 'label', content: new Map([
-                        ["nl-be", "Fysiek bij de spoedafdeling van het ziekenhuis"],
-                        ["fr-be", "Physiquement au service des urgences de l'hôpital"],
-                        ["de-be", "Physisch in der Notaufnahme des Krankenhauses"],
-                        ["en", "In person at the hospital emergency department"],
-                    ]),
-                },
-                { ...ddOptions }
-            ],
-            displayCondition: expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '31')
-        },
-        {
-            key: 'r32', role: 'responseRow', cells: [
-                {
-                    key: 'col0', role: 'label', content: new Map([
-                        ["nl-be", "Fysiek bij de huisartsenwachtpost"],
-                        ["fr-be", "Physiquement au service de garde de médecine générale"],
-                        ["de-be", "Physisch beim Wachtposten des Hausarztes"],
-                        ["en", "In person at the on-call general medical service "],
-                    ]),
-                },
-                { ...ddOptions }
-            ],
-            displayCondition: expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '32')
-        },
-        {
-            key: 'r6', role: 'responseRow', cells: [
-                {
-                    key: 'col0', role: 'label', content: new Map([
-                        ["nl-be", "Telefonisch of via e-mail contact met huisdokter, receptioniste, verpleger/verpleegster"],
-                        ["fr-be", "Seulement par le biais d'un contact téléphonique ou d'un courrier électronique avec le médecin, la réceptionniste, l'infirmier/l'infirmière"],
-                        ["de-be", "Nur telefonisch oder über einem E-Mail-Kontakt mit Hausarzt, Empfangsdame, Krankenpfleger/Krankenschwester"],
-                        ["en", "Only through telephone or email contact with the doctor, receptionist, nurse "],
-                    ]),
-                },
-                { ...ddOptions }
-            ],
-            displayCondition: expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '6')
-        },
-        {
-            key: 'r2', role: 'responseRow', cells: [
-                {
-                    key: 'col0', role: 'label', content: new Map([
-                        ["nl-be", "Ik ben opgenomen in het ziekenhuis"],
-                        ["fr-be", "J'ai été admis(e) à l'hôpital"],
-                        ["de-be", "Ich wurde in das Krankenhaus aufgenommen"],
-                        ["en", "I was admitted into the hospital"],
-                    ]),
-                },
-                { ...ddOptions }
-            ],
-            displayCondition: expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '2')
-        },
-        {
-            key: 'r4', role: 'responseRow', cells: [
-                {
-                    key: 'col0', role: 'label', content: new Map([
-                        ["nl-be", "Andere medische hulp."],
-                        ["fr-be", "Autre aide médicale"],
-                        ["de-be", "andere ärztliche Hilfe"],
-                        ["en", "Another healthcare assistance"],
-                    ]),
-                },
-                { ...ddOptions }
-            ],
-            displayCondition: expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '4')
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
-
-/**
- * WHY VISITED NO MEDICAL SERVICE
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const visitedNoMedicalService = (parentKey: string, keyVisitedMedicalServ?: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Qcov_BE_18';
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Wat zijn de voornaamste redenen dat u geen medische hulp zocht in verband met de gerapporteerde symptomen?"],
-            ["fr-be", "Quelles sont les principales raisons pour lesquelles vous n'avez pas consulté un médecin pour les symptômes signalés ?"],
-            ["de-be", "Was sind die wichtigsten Gründe dafür, dass Sie im Zusammenhang mit den gemeldeten Symptomen keine ärztliche Hilfe aufsuchten?"],
-            ["en", "What is the main reason for which you did not contact any health professional for the symptoms you declared today?"],
-        ]))
-    );
-
-    // CONDITION
-    editor.setCondition(
-        expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '0')
-    );
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Waarom vragen we dit?"],
-                    ["fr-be", "Pourquoi posons-nous cette question ?"],
-                    ["de-be", "Warum fragen wir das?"],
-                    ["en", "Why are we asking this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Om uit te zoeken waarom sommige mensen geen medische hulp zoeken."],
-                    ["fr-be", "Pour savoir pourquoi certaines personnes ne consultent pas un médecin."],
-                    ["de-be", "Um festzustellen, warum manche Menschen keine ärztliche Hilfe aufsuchen."],
-                    ["en", "To understand why some people do not consult a doctor."],
-                ]),
-                style: [{ key: 'variant', value: 'p' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Hoe moet ik deze vraag beantwoorden?"],
-                    ["fr-be", "Comment dois-je répondre à cette question ?"],
-                    ["de-be", "Wie soll ich diese Frage beantworten?"],
-                    ["en", "How should I answer this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Meerdere antwoorden mogelijk."],
-                    ["fr-be", "Plusieurs réponses sont possibles."],
-                    ["de-be", "Mehrere Antworten möglich."],
-                    ["en", "Multiple answers are possible."],
-                ]),
-            },
-        ])
-    );
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-2' }],
-        content: generateLocStrings(
-            new Map([
-                ['nl-be', 'Meerdere antwoorden mogelijk'],
-                ["fr-be", "Plusieurs réponses sont possibles"],
-                ["de-be", "Mehrere Antworten möglich"],
-                ["en", "Multiple answers possible"],
-            ])),
-    }, rg?.key);
-
-    const rg_inner = initMultipleChoiceGroup(multipleChoiceKey, [
-        {
-            key: '1', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Mijn symptomen kwamen recent op."],
-                ["fr-be", "Mes symptômes sont apparus récemment."],
-                ["de-be", "Meine Symptome traten erst kürzlich auf."],
-                ["en", "My symptoms appeared very recently"],
-            ])
-        },
-        {
-            key: '2', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Mijn symptomen zijn mild."],
-                ["fr-be", "Mes symptômes sont légers."],
-                ["de-be", "Meine Symptome sind mild."],
-                ["en", "My symptoms are mild"],
-            ])
-        },
-        {
-            key: '3', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Ik heb deze symptomen vaker."],
-                ["fr-be", "Je présente régulièrement ces symptômes."],
-                ["de-be", "Ich habe diese Symptome öfter."],
-                ["en", "I have these symptoms often"],
-            ])
-        },
-        {
-            key: '4', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Ik denk te weten wat ik heb en gebruik eigen medicatie."],
-                ["fr-be", "Je pense que je sais ce dont je souffre, et j'ai recours à ma propre médication."],
-                ["de-be", "Ich meine zu wissen, was ich habe, und verwende meine eigene Medikation."],
-                ["en", "I think I know what I have and I use self-medication"],
-            ])
-        },
-        {
-            key: '5', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Ik denk dat er geen goede behandeling is voor mijn ziekte."],
-                ["fr-be", "Je pense qu'il n'existe pas de bon traitement pour ma maladie."],
-                ["de-be", "Ich denke, dass es keine gute Behandlung für meine Krankheit gibt."],
-                ["en", "I think there is no effective treatment for the disease I have"],
-            ])
-        },
-        {
-            key: '6', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Het is te moeilijk om snel een afspraak te verkrijgen."],
-                ["fr-be", "Il est trop difficile d'obtenir rapidement un rendez-vous."],
-                ["de-be", "Es ist zu schwierig, schnell einen Termin zu bekommen."],
-                ["en", "It is too hard to get an appointment quickly"],
-            ])
-        },
-        {
-            key: '7', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Ik heb onvoldoende tijd."],
-                ["fr-be", "Je n'ai pas assez de temps."],
-                ["de-be", "Ich habe nicht genug Zeit."],
-                ["en", "I do not have time"],
-            ])
-        },
-        {
-            key: '8', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Omwille van financiële redenen."],
-                ["fr-be", "Pour des raisons financières."],
-                ["de-be", "Aus finanziellen Gründen."],
-                ["en", "For financial reasons"],
-            ])
-        },
-        {
-            key: '9', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            content: new Map([
-                ["nl-be", "Omwille van angst voor de gevolgen als de dokter vermoedt dat ik COVID-19 heb."],
-                ["fr-be", "Par crainte des conséquences si le médecin me suspecte d'avoir contracté le coronavirus."],
-                ["de-be", "Aufgrund von Angst vor den Folgen, wenn der Arzt vermutet, dass ich COVID-19 habe."],
-                ["en", "For fear of consequences if the doctor thinks I have COVID-19"],
-            ])
-        },
-        {
-            key: '10', role: 'input',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '11'),
-            style: [{ key: 'className', value: 'w-100' }],
-            content: new Map([
-                ["nl-be", "Andere"],
-                ["fr-be", "Autre"],
-                ["de-be", "Andere"],
-                ["en", "Other"],
-            ]),
-            description: new Map([
-                ["nl-be", "Beschrijf hier (optioneel in te vullen)"],
-                ["fr-be", "Veuillez fournir une description ici (facultatif)"],
-                ["de-be", "Beschreiben Sie es hier (optional einzutragen)"],
-                ["en", "Describe here (optional)"],
-            ])
-        },
-        {
-            key: '11', role: 'option',
-            content: new Map([
-                ["nl-be", "Ik weet het niet/verkies niet te antwoorden."],
-                ["fr-be", "Je ne sais pas/je préfère ne pas répondre."],
-                ["de-be", "Ich weiss es nicht/ziehe es vor, nicht zu antworten."],
-                ["en", "I don't know/can't remember"],
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
-
-/**
- * FEAR CONSEQUENCES
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const consFear = (parentKey: string, keyVisitedMedicalServ?: string, keyvisitedNoMedicalService?: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Qcov_BE_18b';
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Kan u aangeven welke gevolgen u vreest?"],
-            ["fr-be", "Pouvez-vous indiquer les conséquences que vous craignez ?"],
-            ["de-be", "Können Sie angeben, welche Folgen Sie fürchten?"],
-            ["en", "Can you specify which consequences you mainly fear? "],
-        ]))
-    );
-
-    // CONDITION
-    editor.setCondition(
-        expWithArgs('and',
-            expWithArgs('responseHasKeysAny', keyVisitedMedicalServ, [responseGroupKey, multipleChoiceKey].join('.'), '0'),
-            expWithArgs('responseHasKeysAny', keyvisitedNoMedicalService, [responseGroupKey, multipleChoiceKey].join('.'), '9')
-        )
-    );
-
-    // INFO POPUP
-    // None
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-2' }],
-        content: generateLocStrings(
-            new Map([
-                ['nl-be', 'Meerdere antwoorden mogelijk'],
-                ["fr-be", "Plusieurs réponses sont possibles"],
-                ["de-be", "Mehrere Antworten möglich"],
-                ["en", "Multiple answers possible"],
-            ])),
-    }, rg?.key);
-
-    const rg_inner = initMultipleChoiceGroup(multipleChoiceKey, [
-        {
-            key: '1', role: 'option',
-            content: new Map([
-                ["nl-be", "Ik moet een test laten uitvoeren."],
-                ["fr-be", "Je dois me faire tester."],
-                ["de-be", "Ich muss einen Test durchführen lassen."],
-                ["en", "Have to do a diagnostic test."],
-            ])
-        },
-        {
-            key: '2', role: 'option',
-            content: new Map([
-                ["nl-be", "Ik moet in quarantaine."],
-                ["fr-be", "Je dois me mettre en quarantaine."],
-                ["de-be", "Ich muss in Quarantäne gehen."],
-                ["en", "Have to be isolated."],
-            ])
-        },
-        {
-            key: '3', role: 'option',
-            content: new Map([
-                ["nl-be", "Ik moet stoppen met werken."],
-                ["fr-be", "Je dois arrêter de travailler."],
-                ["de-be", "Ich muss mit dem Arbeiten aufhören."],
-                ["en", "Have to stop working."],
-            ])
-        },
-        {
-            key: '45', role: 'option',
-            content: new Map([
-                ["nl-be", "Veroordeeld worden door anderen."],
-                ["fr-be", "Le jugement des autres."],
-                ["de-be", "Durch andere verurteilt werden."],
-                ["en", "To be judged by others."],
-            ])
-        },
-        {
-            key: '6', role: 'input',
-            style: [{ key: 'className', value: 'w-100' }],
-            content: new Map([
-                ["nl-be", "Andere"],
-                ["fr-be", "Autre"],
-                ["de-be", "Andere"],
-                ["en", "Other"],
-            ]),
-            description: new Map([
-                ["nl-be", "Beschrijf hier (optioneel in te vullen)"],
-                ["fr-be", "Veuillez fournir une description ici (facultatif)"],
-                ["de-be", "Beschreiben Sie es hier (optional einzutragen)"],
-                ["en", "Describe here (optional)"],
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
 
     // VALIDATIONs
     if (isRequired) {
@@ -2369,142 +1404,6 @@ const inComa = (parentKey: string, keyVisitedMedicalServ: string, isRequired?: b
     return editor.getItem();
 }
 
-/**
- * SYMPTOM IMPLIED COVID-19 TEST PERFORMED
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const SymptomImpliedCovidTest = (parentKey: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Qcov_BE_16'
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Heeft u omwille van uw symptomen een test laten uitvoeren voor COVID-19? "],
-            ["fr-be", "Avez-vous passé un test de dépistage du coronavirus en raison de vos symptômes ?"],
-            ["de-be", "Haben Sie aufgrund Ihrer Symptome einen Test auf COVID-19 durchführen lassen?"],
-            ["en", "Because of your symptoms, did you undergo a test/analyses to know if you have COVID-19?"],
-        ]))
-    );
-
-    // CONDITION
-    // none
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Waarom vragen we dit?"],
-                    ["fr-be", "Pourquoi posons-nous cette question ?"],
-                    ["de-be", "Warum fragen wir das?"],
-                    ["en", "Why are we asking this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "We willen weten voor welke klachten mensen zich laten testen op COVID-19."],
-                    ["fr-be", "Nous voulons savoir sur la base de quelles plaintes les gens se font tester en vue de dépister le coronavirus."],
-                    ["de-be", "Wir möchten wissen, für welche Beschwerden Menschen sich auf COVID-19 testen lassen."],
-                    ["en", "We want to know which complaints lead people to get tested for the coronavirus."],
-                ]),
-                // style: [{ key: 'variant', value: 'p' }],
-            },
-        ])
-    );
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    editor.addExistingResponseComponent({
-        role: 'text',
-        content: generateLocStrings(
-            new Map([
-                ['nl-be', 'Meerdere antwoorden mogelijk'],
-                ["fr-be", "Plusieurs réponses sont possibles"],
-                ["de-be", "Mehrere Antworten möglich"],
-                ["en", "Multiple answers possible"],
-            ])),
-        style: [{ key: 'className', value: 'mb-1' }]
-    }, rg?.key);
-    const rg_inner = initMultipleChoiceGroup(multipleChoiceKey, [
-        // {
-        //     key: '0', role: 'option', content: new Map([
-        //         ["nl-be", "Nee"],
-        //         ["fr-be", "Non"],
-        //         ["de-be", "Nein"],
-        //         ["en", "No"],
-        //     ])
-        // },
-        {
-            key: '1', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '3', '4'),
-            content: new Map([
-                ["nl-be", "Ja, een test uitgevoerd op basis van een wattenstaafje in mijn neus of mond"],
-                ["fr-be", "Oui, un test effectué à l'aide d'un écouvillon dans mon nez ou ma bouche"],
-                ["de-be", "Ja, ein Test wurde mit einem Wattestäbchen in meiner Nase oder Mund durchgeführt"],
-                ["en", "Yes, a test based on a swab in nose or mouth, or a sputum or saliva sample"],
-            ])
-        },
-        {
-            key: '2', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '3', '4'),
-            content: new Map([
-                ["nl-be", "Ja, een bloedtest"],
-                ["fr-be", "Oui, un test sanguin"],
-                ["de-be", "Ja, ein Bluttest"],
-                ["en", "Yes, a serological analysis (screening for antibodies against this virus, from a drop of blood from fingertip or a blood sample)"],
-            ])
-        },
-        {
-            key: '5', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '3', '4'),
-            content: new Map([
-                ["nl-be", "Ja, een sneltest (antigeentest) (met een wattenstaafje in mijn neus of mond, en met een resultaat beschikbaar binnen het uur)"],
-                ["fr-be", "Oui, un test antigénique rapide (sur un écouvillon introduit dans le nez ou la bouche, permettant d'obtenir un résultat en moins d'une heure)"],
-                ["de-be", "Ja, ein Antigen-Schnelltest (mit einem Wattestäbchen in Nase oder Mund mit einem Ergebnis, das in unter einer Stunde verfügbar ist)"],
-                ["en", "Yes, a rapid antigen detection test (on a swab in nose or mouth, with a result available in less than an hour)"],
-            ])
-        },
-        {
-            key: '3', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '1', '2', '4', '5'),
-            content: new Map([
-                ["nl-be", "Nog niet, ik ga binnenkort een test laten uitvoeren"],
-                ["fr-be", "Pas encore, je vais bientôt me faire tester"],
-                ["de-be", "Noch nicht, ich werde in Kürze einen Test durchführen lassen"],
-                ["en", "Not yet, I plan to shortly undergo a test"],
-            ])
-        },
-        {
-            key: '4', role: 'option',
-            disabled: expWithArgs('responseHasKeysAny', editor.getItem().key, responseGroupKey + '.' + multipleChoiceKey, '1', '2', '3', '5'),
-            content: new Map([
-                ["nl-be", "Nee, ik zal geen test laten uitvoeren"],
-                ["fr-be", "Non, je ne me ferai pas tester"],
-                ["de-be", "Nein, ich werde keinen Test durchführen lassen"],
-                ["en", "No, I will not get tested"],
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
 
 /**
  * COVID 19 Personal Habits Changes: likert scale question about changes in personal habits after experiencing covid symptoms
