+++ src/editor-example-generators/surveys.ts
@@ -1,5 +1,6 @@
 import getBelIntake from './belgium/inf-intake';
 import getItalyIntake from './italy/inf-intake';
+import getItalyWeekly from './italy/inf-weekly';
 import getBelWeekly from './belgium/inf-weekly';
 import getBelVaccination from './belgium/inf-vaccination';
 import getExampleIntake from './nl/infectieradar/inf-intake';
@@ -67,6 +68,7 @@ const surveys = [
         instance: 'italy',
         surveys: [
             { name: "intake", survey: getItalyIntake() },
+            { name: "weekly", survey: getItalyWeekly() },
         ],
         languageCodes: [
             'en',
